fix(tests): extract projector-pg helpers to common module; replace internal tracker refs with GH numbers
- Extract test_url/make_pool/new_ctx/setup_tenant/teardown_tenant to
  tests/common/mod.rs; projector-pg/tests/integration.rs drops from
  649 → 564 lines, passing the 600-line CI cap
- Replace internal tracker IDs with GitHub issue/PR numbers in all code comments
  across parse-worker and projector-pg test/source files
